Refactor Scanner
Refactored the Scanner to use a list of Matcher delegates.
This allows the code for recognizing each token to be in one place
rather than two.

Also it allowed us to construct matching code for the very common cases
of a single character or predicate.

Also added a (currently disabled) test for Quasiquote.