Allow Environment to be mutable
It might be interesting to try to enforce immutability,
but we're not implementing a Haskell interpreter, and in
any case, Haskell never supported a REPL because
of the whole problem of what to do about define.

I'd like to include 'define' and making environments mutable is
the simplest and most direct way to accomplish this.