Add support for matching symbols in case lambda
Extend the quotation syntax so that a 'quoted' expression in
a lambda argument list implies an exact match.

This allows us to match against exact symbols as well as exact atoms.

We could go further: unquote could mean "evaluate the enclosed expression".
This would allow us to pattern match against values in the environment
at the time of the lambda instantiation. This is required if we ever
want to pattern match against anything that doesn't have a literal
representation.