Make FCLisp properly tail recursive
Finished the explicit stack based evaluation.

FCLisp now does not depend on the size or "nature" of the .Net runtime stack.
One benefit of this will be that we'll be able to more easily add debugging
facilities and a "stack trace".

Another benefit is that it is now "properly tail recursive" in that
functions written in a properly tail recursive manner only consume O(1)
stack space.

Finally, we will be able to implement call/cc by making the stack "immutable".