Add 'begin' as separate core form
'begin' allows us to execute a sequence of statements and only
pay attention to the last one. Useful for adding debug print statements.

redefined 'define' as a 'raw' version that doesn't have the 'begin'
functionality, then re-implemented the original full 'define' behaviour
as a macro in builtins.lisp

First "real" use of quasiquote and it worked!