Add conventional "define function" syntax
The traditional way to define a function in Scheme and
particularly the SICP book is:

(define (square x) (* x x)) which is shorthand for

(define square (lambda (x) (* x x)))

Extended the 'define' macro to support this usage.

Also, in case of there being multiple bodies, only the
bottom one is used.