Add set! as a builtin
We're done with immutability. If we'd stuck with it, I suppose
we'd have had to start inventing Monads and all that to make
it convenient have no implicit state. But our goals are less
lofty. set! is added. Actually, it's good because I hadn't
realised the difference between set! and define until now:
set! can change *anything* in the current lexical scope, whereas
define only "adds" to the current environment.