Allow each .lisp test file to have a setup
Each .lisp file now has a "setup" element and a "tests" element.
The "setup" allows arbitrary lisp code to be executed once before
all of the tests are run.

This allows us to not have "mscorlib" in builtins.lisp, and also
allows more sophisticated tests to be written at a later point
without requiring more and more logic to be added to builtins.lisp.