Add a unit test showing problem with .Net methods
The .Net methods were implemented by explicitly using an evaluator
to evaluate each argument rather than delegating this to the calling
evaluator.

The problem with this is that it breaks continuations. Unit test gives
an example of how this could be a problem: constructing an argument
for a .Net method call which happens to raise a continuation
only lets the continuation escape to the site of the .Net method call.