Implement instance methods as a macro
"(.method target arg1 arg2 ...)" is now a macro which is expanded into
"(invoke-instance target "method" (list arg1 arg2 ...))"

Doing things this way we no longer require "." to be a special "FExpression".
This solves all of the evaluator problems (that broke continuations
for .Net methods). It could have been solved by changing the "ClrSymbol"
FExpression to obey all the rules - but this way was easier.

This reflection ability is in the "standard" environment - but it's also
possible to create a "sandbox" environment that cannot arbitrarily call
into reflection.

All that's left now is modify how static methods will work.

Simplest way is probably to expose all static methods in an assembly
whenever "ref" is used...