Re-implement expand in Lisp (mostly)
The only part of 'expand' that had to be a builtin was 'unmacro' - the inverse
function of 'macro'. Given that function, the rest of the functionality
can be implemented in terms of eval and normal lisp.

The net result: less C# code, which will be easier to refactor later if
we change the way that the interpreter works.