Make the REPL robust when scanner fails
Typing an unrecognized token was making the entire REPL exit due to perceived 'EOF'.
This is because IEnumerator.MoveNext() implementation of "yield return"
appears to switch to an "EOF" state after an exception is thrown.

Now, we create a new IEnumerator after any exception during scanning has occurred, so that if you
choose to continue "read"ing then you won't get EOF unless it's *actually* the end
of the input stream.