Remove need for '(ref mscorlib)'
Big change to ".net" method syntax.

Symbols containing "." are expanded into a "dot" macro but now with each
component split up. So "System.Console" comes back from the reader as
(dot System Console), and ".Equals" expands to (dot () Equals).

But also: Symbols containing a "/" are also expanded into a "slash" macro, unless
the symbol contains only a slash. So "System.Console/WriteLine" comes back
as (slash (dot System Console) WriteLine).

The net effect now is that we have an unambiguous but convenient way to refer to
types, static methods, and instance methods with needing them to be pre-defined
in the environment.

This leads to a big performance improvement as 'environment' does not need to
be anywhere near as big (previously, '(ref mscorlib)' added hundreds of symbols
to the environment which meant anything defined before them was very expensive to
lookup.

The only downside is "let/cc" and "call/cc" had to be renamed to "let-cc" and "call-cc"
which is not really that bad.