Add test to highlight symbol caching graph
symbol-lookup-cache highlights the pitfals of caching information inside the Symbol datum.
On the one hand, this is the fastest place to look for it. On the other hand, if the S-expr
that we receive is actually a graph rather than a tree in which each node has a different
lexical context, this is dangerous.

This doesn't happen for things read by the reader. In generally it only happens
via macros or 'eval' in which some arbitrary code has 'constructed' a graph for us to evaluate.

The fix will be to 'clone' all such graphs as trees after macro expansion and just before 'eval'.